backend/requirements.txt
pwdlib[argon2]
fastapi
pyjwt
pwdlib
sqlmodel
uvicorn
pydantic
pydantic[email]
python-dotenv
python-multipart
requests
slowapi
pytest
httpx